python-analysis-server/requirements.txt
fastapi
uvicorn
pydantic
librosa
numpy
scipy
google-generativeai
python-dotenv